+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # change log
 
+## `0.0.3`
+
+- add `length` getter
+- swallow error from worker
+- refactor `process` method, `process` should be idempotent
+- refactor `drain` method, use deferred resolve
+
 ## `0.0.2`
 
 - add `drain` method